WIP Shader creation and compilation

+++ Source/Plugins/TeFreeImgImporter/TeFreeImgImporter.cpp
@@ -248,7 +248,7 @@ namespace te
 
         imageFormat = (FREE_IMAGE_FORMAT)findFormat->second;
         
-        uint8_t* data = static_cast<uint8_t*>(te_allocate(size));
+        uint8_t* data = static_cast<uint8_t*>(te_allocate(static_cast<UINT32>(size)));
         memset(data, 0, size);
         file.Read(static_cast<char*>((void*)data), static_cast<std::streamsize>(size));
